+++ api/specs/webserver/scripts/openapi_nih_sparc.py
@@ -8,12 +8,8 @@
 # pylint: disable=unused-variable
 
 
-from fastapi import APIRouter, FastAPI, status
-from fastapi.responses import RedirectResponse
+from fastapi import APIRouter, FastAPI
 from models_library.generics import Envelope
-from models_library.projects import ProjectID
-from models_library.services import ServiceKey, ServiceKeyVersion
-from pydantic import HttpUrl, PositiveInt
 from simcore_service_webserver._meta import API_VTAG
 from simcore_service_webserver.studies_dispatcher._rest_handlers import (
     ServiceGet,
@@ -65,35 +61,6 @@ async def list_default_viewers(file_type: str | None = None):
 
     If file_type is provided, then it filters viewer for that filetype
     """
-
-
-@router.get(
-    "/view",
-    response_class=RedirectResponse,
-    response_description="Opens osparc and starts viewer for selected data",
-    status_code=status.HTTP_302_FOUND,
-    operation_id="get_redirection_to_viewer",
-)
-async def get_redirection_to_viewer(
-    file_type: str,
-    viewer_key: ServiceKey,
-    viewer_version: ServiceKeyVersion,
-    file_size: PositiveInt,
-    download_link: HttpUrl,
-    file_name: str | None = "unknown",
-):
-    """Opens a viewer in osparc for data in the NIH-sparc portal"""
-
-
-@router.get(
-    "/study/{id}",
-    response_class=RedirectResponse,
-    response_description="Opens osparc and opens a copy of publised study",
-    status_code=status.HTTP_302_FOUND,
-    operation_id="get_redirection_to_study_page",
-)
-async def get_redirection_to_study_page(id: ProjectID):
-    """Opens a study published in osparc"""
 
 
 if __name__ == "__main__":
